Made table drop operations safe

+++ db/migrate/20170201161340_drop_chess_piece_table.rb
@@ -1,5 +1,5 @@
 class DropChessPieceTable < ActiveRecord::Migration[5.0]
   def change
-        drop_table :chess_pieces
+    drop_table :chess_pieces if ActiveRecord::Base.connection.table_exists? :chess_pieces
   end
 end
